+++ src/components/BackBtn/BackBtn.stories.ts
@@ -18,6 +18,34 @@ const meta = {
 		hideBackIcon: {
 			control: 'boolean',
 		},
+		locales: {
+			description: 'Traductions',
+			control: false,
+			table: {
+				category: 'props',
+				type: {
+					summary: undefined,
+				},
+				defaultValue: {
+					summary: `Locales`,
+					detail: `{
+	label: 'Retour',
+}`,
+				},
+			},
+		},
+		backgroundColor: {
+			control: { type: 'text' },
+			description: 'Couleur de fond du bouton parmi la palette du thème.',
+			table: {
+				type: {
+					summary: 'string',
+				},
+				defaultValue: {
+					summary: 'white',
+				},
+			},
+		},
 	},
 } as Meta<typeof BackBtn>
 
@@ -61,12 +89,12 @@ export const Default: Story = {
 					color="primary"
 					class="pa-4"
 				>
-				<BackBtn v-bind="args" />
-				</VSheet>
-				<VSheet
-					v-else
-					class="pa-4"
-				>
+					<BackBtn v-bind="args" />
+					</VSheet>
+					<VSheet
+						v-else
+						class="pa-4"
+					>
 					<BackBtn v-bind="args" />
 				</VSheet>
 			`,