+++ src/components/HeaderToolbar/tests/HeaderToolBar.spec.ts
@@ -178,7 +178,7 @@ describe('HeaderToolbar', () => {
 
 		expect(wrapper.vm.highlightMenu).toBe(false)
 
-		const activeSelected = wrapper.find('.sy-input-select > span')
+		const activeSelected = wrapper.find('.sy-header-button')
 		expect(activeSelected.text()).toBe('Professionnel de santé')
 	})
 
@@ -221,7 +221,7 @@ describe('HeaderToolbar', () => {
 
 		await wrapper.vm.$nextTick()
 
-		const activeSelected = wrapper.find('.sy-input-select > span')
+		const activeSelected = wrapper.find('.sy-header-button')
 		expect(activeSelected.text()).toBe('Professionnel de santé')
 
 		await wrapper.vm.handleLink(1)