+++ src/components/PhoneField/PhoneField.stories.ts
@@ -32,6 +32,7 @@ const meta = {
 		isValidatedOnBlur: { control: 'boolean' },
 		displayAsterisk: { control: 'boolean' },
 		disableErrorHandling: { control: 'boolean' },
+		disabled: { control: 'boolean' },
 	},
 } satisfies Meta<typeof PhoneField>
 
@@ -89,6 +90,9 @@ export const Default: Story = {
 		customIndicatifs: [],
 		useCustomIndicatifsOnly: false,
 		isValidatedOnBlur: true,
+		bgColor: undefined,
+		readOnly: false,
+		disabled: false,
 	},
 	render: (args) => {
 		return {
@@ -109,6 +113,9 @@ export const Default: Story = {
 					:customIndicatifs="args.customIndicatifs"
 					:useCustomIndicatifsOnly="args.useCustomIndicatifsOnly"
 					:isValidatedOnBlur="args.isValidatedOnBlur"
+					:bg-color="args.bgColor"
+					:readOnly="args.readOnly"
+					:disabled="args.disabled"
 				/>
 				</div>
 				<br><br><br><br><br>
@@ -167,6 +174,9 @@ export const Required: Story = {
 		customIndicatifs: [],
 		useCustomIndicatifsOnly: false,
 		isValidatedOnBlur: true,
+		bgColor: undefined,
+		readOnly: false,
+		disabled: false,
 	},
 	render: (args) => {
 		return {
@@ -187,6 +197,9 @@ export const Required: Story = {
      :customIndicatifs="args.customIndicatifs"
      :useCustomIndicatifsOnly="args.useCustomIndicatifsOnly"
      :isValidatedOnBlur="args.isValidatedOnBlur"
+     :bg-color="args.bgColor"
+	 :readOnly="args.readOnly"
+	 :disabled="args.disabled"
     />
 				</div>
    `,
@@ -202,6 +215,7 @@ export const RequiredWithAsterisk: Story = {
 		...Default.args,
 		required: true,
 		displayAsterisk: true,
+		bgColor: undefined,
 	},
 	parameters: {
 		docs: {
@@ -255,6 +269,9 @@ const phoneNumber = ref('')
 						:useCustomIndicatifsOnly="args.useCustomIndicatifsOnly"
 						:isValidatedOnBlur="args.isValidatedOnBlur"
 						:display-asterisk="args.displayAsterisk"
+						:readOnly="args.readOnly"
+						:disabled="args.disabled"
+						:bg-color="args.bgColor"
 					/>
 				</div>
 			`,
@@ -322,6 +339,9 @@ export const CustomIndicatifs: Story = {
 		],
 		useCustomIndicatifsOnly: true,
 		isValidatedOnBlur: true,
+		bgColor: undefined,
+		readOnly: false,
+		disabled: false,
 	},
 	render: (args) => {
 		return {
@@ -342,6 +362,9 @@ export const CustomIndicatifs: Story = {
      :customIndicatifs="args.customIndicatifs"
      :useCustomIndicatifsOnly="args.useCustomIndicatifsOnly"
      :isValidatedOnBlur="args.isValidatedOnBlur"
+	 :readOnly="args.readOnly"
+	 :disabled="args.disabled"
+     :bg-color="args.bgColor"
     />
 				</div>
    `,
@@ -399,6 +422,9 @@ export const NotValidatedOnBlur: Story = {
 		customIndicatifs: [],
 		useCustomIndicatifsOnly: false,
 		isValidatedOnBlur: false,
+		readOnly: false,
+		disabled: false,
+		bgColor: undefined,
 	},
 	render: (args) => {
 		return {
@@ -419,6 +445,9 @@ export const NotValidatedOnBlur: Story = {
      :customIndicatifs="args.customIndicatifs"
      :useCustomIndicatifsOnly="args.useCustomIndicatifsOnly"
      :isValidatedOnBlur="args.isValidatedOnBlur"
+	 :readOnly="args.readOnly"
+	 :disabled="args.disabled"
+     :bg-color="args.bgColor"
     />
 				</div>
    `,
@@ -476,6 +505,9 @@ export const DisplayFormatCode: Story = {
 		customIndicatifs: [],
 		useCustomIndicatifsOnly: false,
 		isValidatedOnBlur: false,
+		readOnly: false,
+		disabled: false,
+		bgColor: undefined,
 	},
 	render: (args) => {
 		return {
@@ -496,6 +528,9 @@ export const DisplayFormatCode: Story = {
      :customIndicatifs="args.customIndicatifs"
      :useCustomIndicatifsOnly="args.useCustomIndicatifsOnly"
      :isValidatedOnBlur="args.isValidatedOnBlur"
+	 :readOnly="args.readOnly"
+	 :disabled="args.disabled"
+     :bg-color="args.bgColor"
     />
 				</div>
    `,
@@ -553,6 +588,9 @@ export const DisplayFormatCodeAbbreviation: Story = {
 		customIndicatifs: [],
 		useCustomIndicatifsOnly: false,
 		isValidatedOnBlur: false,
+		readOnly: false,
+		disabled: false,
+		bgColor: undefined,
 	},
 	render: (args) => {
 		return {
@@ -573,6 +611,9 @@ export const DisplayFormatCodeAbbreviation: Story = {
      :customIndicatifs="args.customIndicatifs"
      :useCustomIndicatifsOnly="args.useCustomIndicatifsOnly"
      :isValidatedOnBlur="args.isValidatedOnBlur"
+	 :readOnly="args.readOnly"
+	 :disabled="args.disabled"
+     :bg-color="args.bgColor"
     />
 				</div>
    `,
@@ -630,6 +671,9 @@ export const DisplayFormatCodeCountry: Story = {
 		customIndicatifs: [],
 		useCustomIndicatifsOnly: false,
 		isValidatedOnBlur: false,
+		readOnly: false,
+		disabled: false,
+		bgColor: undefined,
 	},
 	render: (args) => {
 		return {
@@ -650,6 +694,9 @@ export const DisplayFormatCodeCountry: Story = {
      :customIndicatifs="args.customIndicatifs"
      :useCustomIndicatifsOnly="args.useCustomIndicatifsOnly"
      :isValidatedOnBlur="args.isValidatedOnBlur"
+	 :readOnly="args.readOnly"
+	 :disabled="args.disabled"
+     :bg-color="args.bgColor"
     />
 				</div>
    `,
@@ -707,6 +754,9 @@ export const DisplayFormatCountry: Story = {
 		customIndicatifs: [],
 		useCustomIndicatifsOnly: false,
 		isValidatedOnBlur: false,
+		readOnly: false,
+		disabled: false,
+		bgColor: undefined,
 	},
 	render: (args) => {
 		return {
@@ -727,6 +777,9 @@ export const DisplayFormatCountry: Story = {
      :customIndicatifs="args.customIndicatifs"
      :useCustomIndicatifsOnly="args.useCustomIndicatifsOnly"
      :isValidatedOnBlur="args.isValidatedOnBlur"
+	 :readOnly="args.readOnly"
+	 :disabled="args.disabled"
+     :bg-color="args.bgColor"
     />
 				</div>
    `,
@@ -784,6 +837,9 @@ export const DisplayFormatAbbreviation: Story = {
 		customIndicatifs: [],
 		useCustomIndicatifsOnly: false,
 		isValidatedOnBlur: false,
+		readOnly: false,
+		disabled: false,
+		bgColor: undefined,
 	},
 	render: (args) => {
 		return {
@@ -804,6 +860,9 @@ export const DisplayFormatAbbreviation: Story = {
      :customIndicatifs="args.customIndicatifs"
      :useCustomIndicatifsOnly="args.useCustomIndicatifsOnly"
      :isValidatedOnBlur="args.isValidatedOnBlur"
+	 :readOnly="args.readOnly"
+	 :disabled="args.disabled"
+     :bg-color="args.bgColor"
     />
 							</div>
    `,
@@ -867,6 +926,8 @@ export const DisplayModels: Story = {
 		customIndicatifs: [],
 		useCustomIndicatifsOnly: false,
 		isValidatedOnBlur: false,
+		readOnly: false,
+		bgColor: undefined,
 	},
 	render: (args) => {
 		return {
@@ -891,6 +952,9 @@ export const DisplayModels: Story = {
 						:customIndicatifs="args.customIndicatifs"
 						:useCustomIndicatifsOnly="args.useCustomIndicatifsOnly"
 						:isValidatedOnBlur="args.isValidatedOnBlur"
+						:readOnly="args.readOnly"
+						:disabled="args.disabled"
+						:bg-color="args.bgColor"
 					/>
 				</div>
 			`,
@@ -951,6 +1015,9 @@ export const DisabledErrorHandling: Story = {
 		useCustomIndicatifsOnly: false,
 		isValidatedOnBlur: true,
 		disableErrorHandling: true,
+		readOnly: false,
+		disabled: false,
+		bgColor: undefined,
 	},
 	render: (args) => {
 		return {
@@ -974,6 +1041,9 @@ export const DisabledErrorHandling: Story = {
 						:useCustomIndicatifsOnly="args.useCustomIndicatifsOnly"
 						:isValidatedOnBlur="args.isValidatedOnBlur"
 						:disableErrorHandling="args.disableErrorHandling"
+						:readOnly="args.readOnly"
+						:disabled="args.disabled"
+						:bg-color="args.bgColor"
 					/>
 					<div class="mt-6">
 						<h3>Comparaison avec gestion des erreurs activée</h3>
@@ -990,6 +1060,9 @@ export const DisabledErrorHandling: Story = {
 							:useCustomIndicatifsOnly="args.useCustomIndicatifsOnly"
 							:isValidatedOnBlur="args.isValidatedOnBlur"
 							:disableErrorHandling="false"
+							:readOnly="args.readOnly"
+							:disabled="args.disabled"
+							:bg-color="args.bgColor"
 						/>
 					</div>
 				</div>