+++ src/components/PhoneField/PhoneField.vue
@@ -33,6 +33,9 @@
 		isValidatedOnBlur: { type: Boolean, default: true },
 		displayAsterisk: { type: Boolean, default: false },
 		disableErrorHandling: { type: Boolean, default: false },
+		bgColor: { type: String, default: undefined },
+		readOnly: { type: Boolean, default: false },
+		disabled: { type: Boolean, default: false },
 	})
 
 	const emit = defineEmits(['update:modelValue', 'update:selectedDialCode', 'change'])
@@ -186,6 +189,9 @@
 			:display-asterisk="props.displayAsterisk"
 			:disable-error-handling="props.disableErrorHandling"
 			:return-object="true"
+			:bg-color="props.bgColor"
+			:read-only="props.readOnly"
+			:disabled="props.disabled"
 			class="custom-select"
 			text-key="displayText"
 			value-key="code"
@@ -203,6 +209,8 @@
 			:success-messages="successes"
 			:variant="outlined ? 'outlined' : 'underlined'"
 			:display-asterisk="props.displayAsterisk"
+			:bg-color="props.bgColor"
+			:disabled="props.disabled"
 			:class="{
 				'phone-field': true,
 				'error-field': hasError,