+++ src/components/SubHeader/SubHeader.stories.ts
@@ -80,7 +80,7 @@ const meta = {
 					location: 'end center',
 					offset: 16,
 					zIndex: 8,
-					contentClass: 'vd-copy-tooltip-menu text-white text-body-2 ml-2',
+					contentClass: 'sy-copy-tooltip-menu text-white text-body-2 ml-2',
 				},
 				btn: {
 					icon: true,