+++ vite.config.ts
@@ -65,9 +65,9 @@ function generateVuetifyGlobals() {
 export default defineConfig({
 	plugins: [
 		dts({
-			exclude: ['**/*.stories.ts'],
+			exclude: ['**/*.stories.ts', '**/*.spec.ts'],
 			entryRoot: 'src',
-			outDir: 'dist/src',
+			outDir: 'dist',
 			tsconfigPath: 'tsconfig.app.json',
 			rollupTypes: false,
 			insertTypesEntry: true,
@@ -127,6 +127,9 @@ export default defineConfig({
 				inline: ['vuetify'],
 			},
 		},
+		snapshotSerializers: [
+			'./node_modules/vue3-snapshot-serializer/index.js',
+		],
 		coverage: {
 			enabled: true,
 			provider: 'v8',